Release 0.3 bug "NoSuchFieldError: No static field listView1 of type I in class Lcom/disdemo/R$id; or its superclasses"
