+++ App1/App1/HomeActivity.cs
@@ -10,20 +10,18 @@ using Android.Runtime;
 using Android.Views;
 using Android.Widget;
 using Android.Support.V7.App;
-using GR.Net.Maroulis.Library;
+using App1;
 
 namespace UGYM
 {
     [Activity(Label = "HomeActivity", Theme ="@style/Theme.AppCompat.Light.NoActionBar")]
     public class HomeActivity : AppCompatActivity
     {
-        int count = 0;
 
         protected override void OnCreate(Bundle savedInstanceState)
         {
             base.OnCreate(savedInstanceState);
-
-            SetContentView(Resource.Layout.Home);
+            SetContentView(Resource.Layout.Login);
         }
     }
 }
