+++ App1/App1/Resources/layout/Splash.axml
@@ -8,6 +8,6 @@
         android:textSize="20sp"
         android:layout_centerInParent="true"
         android:layout_width="wrap_content"
-        android:layout_height="match_parent"
+        android:layout_height="wrap_content"
         android:layout_centerHorizontal="true" />
 </RelativeLayout>
