+++ App1/App1/UGYM_app.csproj
@@ -90,7 +90,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="HomeActivity.cs" />
-    <Compile Include="SplashScreen.cs" />
+    <Compile Include="SplashActivity.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
@@ -142,7 +142,7 @@
     <AndroidResource Include="Resources\drawable\splash_logo" />
   </ItemGroup>
   <ItemGroup>
-    <AndroidResource Include="Resources\drawable\logo-nuevo.png" />
+    <AndroidResource Include="Resources\drawable\logo_nuevo.png" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
